backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdfplumber
youtube-transcript-api
google-generativeai